requirements.txt
python-telegram-bot>=20,<22
python-dotenv>=1.0,<2
openpyxl>=3.1,<4
pillow
numpy
imageio
imageio-ffmpeg
